packages/source/ceilometer/requirements.txt
d2to1>=0.2.10,<0.3
pbr>=0.5,<0.6
WebOb>=1.2
kombu>2.4.7
iso8601
argparse
SQLAlchemy>=0.7,<=0.7.99
sqlalchemy-migrate>=0.7.2
pymongo>=2.2
eventlet
anyjson>=0.3.3
Flask==0.9
pecan>=0.2.0
stevedore>=0.9
msgpack-python
python-glanceclient
python-novaclient>=2.6.10
python-keystoneclient>=0.2.3
python-swiftclient
lxml
requests>=1.1,<1.2.1
wsme>=0.5b2
pyyaml
oslo.config>=1.1.0
happybase>=0.4
